resume-qa/requirement.txt
flask
llama-index
openai
python-dotenv
PyMuPDF  # For reading PDFs
